+++ packages/extensions/dataflow/src/lnode/lnode-element.test.ts
@@ -1,8 +1,8 @@
 import { describe, it, expect } from 'vitest'
 import { render } from 'vitest-browser-vue'
-import dataflowNode from '../components/dataflow-node.vue'
-import { getLNodeLabel } from '@/types/lnode'
-import type { LNode } from '@/types/lnode'
+import dataflowNode from './lnode-element.vue'
+import { getLNodeLabel } from '@/lnode/lnode'
+import type { LNode } from '@/lnode/lnode'
 
 // Dummy data
 const dummyLNodes: LNode[] = [
@@ -30,11 +30,15 @@ const dummyLNodes: LNode[] = [
 
 describe('dataflow-node', () => {
 	it('renders select options for LNodes and updates displayed label when selected', async () => {
+		// Arrange
+
+		// Act
 		// Render component with dummy lnodes
 		const { container } = render(dataflowNode, {
 			props: { lnodes: dummyLNodes, type: 'input', activeLNodeId: '' },
 		})
 
+		// Assert
 		// Get the select element
 		const select: HTMLSelectElement = container.querySelector('select')!
 		expect(select).toBeTruthy()
@@ -43,14 +47,24 @@ describe('dataflow-node', () => {
 		const options = Array.from(select.options)
 		expect(options.some((opt) => opt.textContent === getLNodeLabel(dummyLNodes[0]))).toBe(true)
 		expect(options.some((opt) => opt.textContent === getLNodeLabel(dummyLNodes[1]))).toBe(true)
+	})
 
+	it('selecting another LNode shows its label', async () => {
+		// Arrange
+		const { container } = render(dataflowNode, {
+			props: { lnodes: dummyLNodes, type: 'input', activeLNodeId: '' },
+		})
+
+		// Act
+		// Render component with dummy lnodes
+		const select: HTMLSelectElement = container.querySelector('select')!
 		// Simulate selecting the second option
 		select.value = dummyLNodes[1].id
 		select.dispatchEvent(new Event('change'))
-
-		// Headline should display selected LNode label
 		await new Promise((r) => setTimeout(r, 0))
 
+		// Assert
+		// Headline should display selected LNode label
 		expect(container.textContent).toContain(getLNodeLabel(dummyLNodes[1]))
 	})
 })
